craated abstract split assigner and bounded version
