+++ flink-connector-bigquery/src/test/java/com/google/cloud/flink/bigquery/source/split/BigQuerySourceSplitAssignerTest.java
@@ -44,10 +44,10 @@ public class BigQuerySourceSplitAssignerTest {
     public void testAssignment() {
         // initialize the assigner with default options since we are faking the bigquery services
         BigQuerySourceSplitAssigner assigner =
-                new BigQuerySourceSplitAssigner(
+                BigQuerySourceSplitAssigner.createBounded(
                         this.readOptions, BigQuerySourceEnumState.initialState());
         // request the retrieval of the bigquery table info
-        assigner.open();
+        assigner.openAndDiscoverSplits();
 
         // should retrieve the first split representing the firt stream
         Optional<BigQuerySourceSplit> maybeSplit = assigner.getNext();
